% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddGCS.R
\name{AddGCS}
\alias{AddGCS}
\title{Add combined Glascow Coma Scale function
This function adds a new column for combined GCS using the motoric, verbal and eye components of GSC.}
\usage{
AddGCS(study_data)
}
\arguments{
\item{study_data}{The study data frame.}
}
\description{
Add combined Glascow Coma Scale function
This function adds a new column for combined GCS using the motoric, verbal and eye components of GSC.
}
