% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddISS.R
\name{AddISS}
\alias{AddISS}
\title{Add iss15 function
This function uses iss to create a binary column of iss > 15}
\usage{
AddISS(study_data)
}
\arguments{
\item{study_data}{The study dataframe.}
}
\description{
Add iss15 function
This function uses iss to create a binary column of iss > 15
}
