% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddMajorS24h.R
\name{AddMajorS24h}
\alias{AddMajorS24h}
\title{Create major surgery < 24 hours function
This function create a new column in the data set for major surgery within 24 hours.}
\usage{
AddMajorS24h(study_data)
}
\arguments{
\item{study_data}{The study data frame.}
}
\description{
Create major surgery < 24 hours function
This function create a new column in the data set for major surgery within 24 hours.
}
