% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddMissingInidcatorVariables.R
\name{AddMissingIndicatorVariables}
\alias{AddMissingIndicatorVariables}
\title{Add missing indicator variables function}
\usage{
AddMissingIndicatorVariables(study.sample, outcome.name = "composite",
  features = NULL, excluded.columns = c("tc"))
}
\arguments{
\item{study.sample}{Data Frame. The study sample. No default}

\item{outcome.name}{String. The name of outcome variable. Defaults to "composite"}

\item{features}{Character vector. The names of the feature variables. Defaults to NULL, in which case all columns except outcome.name and those potentially listed in excluded.columns are treated as features.}

\item{excluded.columns}{Character vector. The names of columns that should not be treated as features. Defaults to NULL and is then ignored. Is also ignored if features is specified.}
}
\description{
Creates missing indicator variables for each predictor variable (feature) with missing values
}
