% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApplyExclusionCriteria.R
\name{ApplyExclusionCriteria}
\alias{ApplyExclusionCriteria}
\title{Apply exclusion criteria function}
\usage{
ApplyExclusionCriteria(study_data)
}
\arguments{
\item{study_data}{The study data as a data frame. No default}
}
\description{
This function applies the study exclusion criteria and saves the results into
the parent environments results list
}
