% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinModels.R
\name{BinModels}
\alias{BinModels}
\title{Bin models function}
\usage{
BinModels(predictions, outcome, n_cores, grid, return_cps = TRUE,
  gridsearch_parallel = FALSE, is_sample = TRUE, maximise = FALSE)
}
\arguments{
\item{predictions}{A numeric vector of predicted probabilites. No default.}

\item{outcome}{A numeric vector of 0 and 1. No default.}

\item{n_cores}{Number of cores to be used in parallel gridsearch. Passed to SupaLarna::gridsearch.breaks. As integer. Defaults to 2 (in gridsearch.breaks)}

\item{grid}{All values to be combnd in gridsearch.breaks, i.e. used to find optimal cutpoints. No default.}

\item{return_cps}{Logical. Function returns model cut_points if TRUE. Defaults to TRUE.}

\item{gridsearch_parallel}{Logical. If TRUE the gridsearch is performed in parallel. Defaults to FALSE.}

\item{is_sample}{Logical. If TRUE, only a tenth of possible cut points is searched. Defaults to TRUE.}

\item{maximise}{Logical. If TRUE, grid search maximises performance metric. Passed to SupaLarna::gridsearch.breaks. Defaults to TRUE.}
}
\description{
This function bins model predictions using SupaLarnas gridsearch.breaks function.
}
