% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinModelVariables.R
\name{BinModelVariables}
\alias{BinModelVariables}
\title{Bin model variables into scores}
\usage{
BinModelVariables(data, model_variables, cut_points, levels)
}
\arguments{
\item{data}{The study data frame. No default.}

\item{model_variables}{The variables to be included in the model. As character vector. No default.}

\item{cut_points}{The cut points used to bin the model variables. As list of numeric vectors. No default.}

\item{levels}{The scores to be assigned to each bin. As list of character vectors. No default.}
}
\description{
This function bins model variables according to cutpoints.
}
