% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootstrapStatistics.R
\name{BootstrapStatistics}
\alias{BootstrapStatistics}
\title{Bootstrap Statistics}
\usage{
BootstrapStatistics(f, log = TRUE, clean.start = TRUE, ...)
}
\arguments{
\item{f}{Function. A function for the statistic argument in boot::boot. No default.}

\item{log}{Logical. If TRUE progress is logged to a logfile. Defaults to TRUE}

\item{clean.start}{Logical. If TRUE the predictions directory and all files in it are removed before saving new stuff there. Defaults to FALSE.}

\item{...}{Arguments for f.}
}
\description{
Wrapper for boot::boot. For the log and clean.start arguments.
}
