% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CollapseMoi.R
\name{CollapseMoi}
\alias{CollapseMoi}
\title{Collapse mechanism of injury function}
\usage{
CollapseMoi(study_data)
}
\arguments{
\item{study_data}{The study data as a data frame. No default}
}
\description{
Takes the main data as input and generates a new mechanism of injury variable with overall levels instead of codes.
}
