% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeAucAndNri.R
\name{ComputeAucAndNri}
\alias{ComputeAucAndNri}
\title{Compute Auc And Nri}
\usage{
ComputeAucAndNri(data, indices, boot = FALSE, log = TRUE,
  clean.start = TRUE, ...)
}
\arguments{
\item{data}{Data frame. The study data. No default.}

\item{indices}{Allows the boot function to pick indices.}

\item{boot}{Logical. If TRUE the sample is treated as a bootstrap sample. Then, indices are used by the boot package to resample the sample, and progress is logged as a bootstrap. Defaults to FALSE}

\item{log}{Logical. If TRUE progress is logged to a logfile. Defaults to TRUE}

\item{clean.start}{Logical. If TRUE logfile is removed and new information is logged. If FALSE information is appended to the logfile. Defaults to FALSE.}
}
\description{
Fit the model to the training data, predict on the test data, and evaluate the predictions.
}
