% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeConfidenceIntervals.R
\name{ComputeConfidenceIntervals}
\alias{ComputeConfidenceIntervals}
\title{Compute Confidence Intervals}
\usage{
ComputeConfidenceIntervals(boot.object, ...)
}
\arguments{
\item{boot.object}{boot class. The boot object. No default.}

\item{...}{Additional arguments for boot::boot.ci.}
}
\description{
Generates confidence intervals around the estimates given in boot.object.
}
