% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeComposite.R
\name{CreateComposite}
\alias{CreateComposite}
\title{Create composite outcome function
This function creates a new binary variable composite outcome}
\usage{
CreateComposite(study_data)
}
\arguments{
\item{study_data}{The study data frame.}
}
\description{
Create composite outcome function
This function creates a new binary variable composite outcome
}
