% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateTable1.R
\name{CreateTable1}
\alias{CreateTable1}
\title{Create Table 1 function
This function creates a table of descriptive statistics for the patient cohort.}
\usage{
CreateTable1(study_data)
}
\arguments{
\item{study_data}{The study data frame.}
}
\description{
Create Table 1 function
This function creates a table of descriptive statistics for the patient cohort.
}
