% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoMedianImputation.R
\name{DoMedianImputation}
\alias{DoMedianImputation}
\title{Do median imputation function}
\usage{
DoMedianImputation(study.sample, missing.indicator = "_missing")
}
\arguments{
\item{study.sample}{Data frame. The study sample. No default.}

\item{missing.indicator}{Character vector of length 1. Used to idenfity variables with missing data to be imputed. Defaults to "_missing"}
}
\description{
Performs median imputation on variables with missing data
}
