% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluateReclassification.R
\name{EvaluateReclassification}
\alias{EvaluateReclassification}
\title{EvaluateReclassification}
\usage{
EvaluateReclassification(current.model.predictions, new.model.predictions,
  outcome.vector, reverse = FALSE)
}
\arguments{
\item{current.model.predictions}{Numeric vector. The risk groups that the current model predicted for subjects. No default.}

\item{new.model.predictions}{Numeric vector. The risk groups that the new model predicted for subjects. No default.}

\item{outcome.vector}{Numeric vector. The outcome of the subjects. No default.}

\item{reverse}{Logical. If TRUE model risk group levels are reversed, i.e. levels 1,2,3,4 are reversed to 4,3,2,1 - respectively. Defaults to FALSE.}
}
\description{
Calculates the NRI.
}
