% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluateWithROCR.R
\name{EvaluateWithRocr}
\alias{EvaluateWithRocr}
\title{Evaluate With Rocr}
\usage{
EvaluateWithRocr(predictions, outcome, measure = "auc",
  reverse = FALSE)
}
\arguments{
\item{predictions}{Numeric vector. Model predictions. No default.}

\item{outcome}{Numeric vector. The outcome of interest. No default.}

\item{measure}{Character vector of length 1. The ROCR measure to estimate. Defaults to "auc".}

\item{reverse}{Logical vector of length 1. If TRUE the functions returns the 1 - AUC rather than AUC. Defaults to FALSE.}
}
\description{
Evaluate the classifier performance using ROCR.
}
