% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FormattingXtable.R
\name{formatting.xtable}
\alias{formatting.xtable}
\title{Format xtable function}
\usage{
formatting.xtable(the_xtable, star_or_caption = NULL,
  table_notes = NULL)
}
\arguments{
\item{the_xtable}{As xtable print. No default.}

\item{star_or_caption}{String. Whether to place comment as footnote or as caption. Accepted strings are "caption" and "star_caption". No Default.}

\item{table_notes}{String. The string to be put under the table as note. Not dynamic, only one note allowed. Defaults to NULL.}
}
\description{
This function takes a xtable print and formats.
}
