% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateFlowchartVect.R
\name{GenerateFlowchartVect}
\alias{GenerateFlowchartVect}
\title{Generate flowchart vector function}
\usage{
GenerateFlowchartVect(results, node_text, exclusion_text,
  results_lst = NULL)
}
\arguments{
\item{results}{Numeric vector or numeric list. Results vector or list. No default.}

\item{node_text}{Character vector. Node text for each of ns in results. In the order in which ns is presented in the results list. No default.}

\item{exclusion_text}{Character vector. Length of node_text - 1. Node text for exclusion nodes. In the order in which ns is presenter in the results list. No default.}

\item{results_lst}{List. Results list, if such a list exist. Defaults to NULL.}
}
\description{
This function merges ns from an input result vector or list, and generates vector for flowchart tikz code. Number of excluded subjects are generated automatically.
}
