% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateModelPredictions.R
\name{GenerateModelPredictions}
\alias{GenerateModelPredictions}
\title{Generate model predictions function}
\usage{
GenerateModelPredictions(data, n_cores, return_cps = FALSE,
  log = FALSE, boot = FALSE, write_to_disk = FALSE,
  clean_start = FALSE, gridsearch_parallel = FALSE, is_sample = TRUE)
}
\arguments{
\item{data}{The study data frame. No default}

\item{n_cores}{Number of cores to be used in parallel gridsearch. Passed to bin.models (which, in turn, passes to SupaLarna::gridsearch.breaks). As integer. Defaults to 2 (in gridsearch.breaks)}

\item{return_cps}{Logical. Function returns model cut_points if TRUE. Passed to bin.models. Defaults to TRUE.}

\item{log}{Logical. If TRUE progress is logged in logfile. Defaults to FALSE.}

\item{boot}{Logical. Affects only what is printed to logfile. If TRUE prepped_sample is assumed to be a bootstrap sample. Defaults to FALSE.}

\item{write_to_disk}{Logical. If TRUE the prediction data is saved as RDS to disk. Defaults to FALSE.}

\item{clean_start}{Logical. If TRUE the predictions directory and all files in it are removed before saving new stuff there. Defaults to FALSE.}

\item{gridsearch_parallel}{Logical. Passed to bin.models (which, in turn, passes to SupaLarnas gridsearch.breaks). If TRUE the gridsearch is performed in parallel. Defaults to FALSE.}

\item{is_sample}{Logical. Passed to bin.models. If TRUE, only a tenth of possible cut points is searched. Defaults to TRUE.}

\item{clean_start}{Logical. If TRUE the predictions directory and all files in it are removed before saving new stuff there. Defaults to FALSE.}
}
\description{
This function bins model predictions and converts them from character labels to numeric labels.
}
