% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDataDictionary.R
\name{GetDataDictionary}
\alias{GetDataDictionary}
\title{Get data dictionary}
\usage{
GetDataDictionary(path = "extdata/", file_name = "data_dictionary.csv")
}
\arguments{
\item{path}{The path to the data dictionary. Defaults to "extdata/".}

\item{file_name}{The name of the data dictionary file. Defaults to "data_dictionary.csv".}
}
\description{
This function reads the data dictionary from a .csv file and returns it as a list
}
