% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GridSearchBreaks.R
\name{GridsearchBreaks}
\alias{GridsearchBreaks}
\title{Gridsearch breaks function}
\usage{
GridsearchBreaks(predictions, outcome, grid = seq(0.01, 0.99, 0.01),
  n.breaks = 3, loss.function = "auc", sample = TRUE,
  parallel = FALSE, n.cores = NULL, save.plot = FALSE,
  verbose = FALSE, maximise = TRUE)
}
\arguments{
\item{predictions}{Numeric vector. Predicted probabilites. No default.}

\item{outcome}{Numeric vecotr. The outcome of interest. No default.}

\item{grid}{Numeric vector. All values to be combined and searched. Defaults to seq(0.01, 0.99, 0.01).}

\item{n.breaks}{Numeric vector of length 1. The number of breaks in the predicted probabilites, i.e. cut-points. Defaults to 3.}

\item{loss.function}{Character vector of length 1. The name of the loss metric, e.g. auc. For now any performance measure included in ROCR is accepted. Defaults to "auc".}

\item{sample}{Logical. If TRUE loss is only calculated for 10% of possible breaks. Defaults to TRUE.}

\item{parallel}{Logical. If TRUE the gridsearch is done in parallel. Defaults to FALSE.}

\item{n.cores}{Numeric vector of length 1. The number of cores to use for parallel computing. Defaults to NULL.}

\item{save.plot}{Logical. If TRUE a wireframe plot is saved to disk. Defaults to FALSE.}

\item{verbose}{Logical. If TRUE then information is printed as the gridsearch runs. Defaults to FALSE. Currently ignored.}

\item{maximise}{Logical. If TRUE, grid search maximizes performance metric. Defaults to TRUE.}
}
\description{
Performs a gridsearch to identify breaks to optimise a provided loss function
}
