% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportData.R
\name{ImportData}
\alias{ImportData}
\title{The Import Data Function}
\usage{
ImportData(file.name, path)
}
\arguments{
\item{file.name}{is the name of your file while path is the path to reach it.
Don't forget to enclose the names as character strings in " ".}
}
\description{
This function imports study data from a csv file.
}
