% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitiateFlowchart.R
\name{InitiateFlowchart}
\alias{InitiateFlowchart}
\title{Initiate Flowchart}
\usage{
InitiateFlowchart(results, ...)
}
\arguments{
\item{results}{List. The results list. No default.}

\item{...}{Args for the bengaltiger::CreateFlowchart.}
}
\description{
Wrapper for the bengaltiger::CreateFlowchart to make the code cleaner
}
