% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeAndSaveXtable.R
\name{make.and.save.xtable}
\alias{make.and.save.xtable}
\title{Make xtable function}
\usage{
make.and.save.xtable(table_data, file_name,
  caption = "Table of Estimates.", label = "analysis_table",
  star_substring = NULL, table_notes = NULL, ...)
}
\arguments{
\item{file_name}{String. File name. No default.}

\item{caption}{String. Caption for table. Defaults to "Table of estimates".}

\item{label}{String. Latex label. Defaults to "analysis_table".}

\item{star_substring}{Character vector. The substrings of xtable to be commented under the table. Defaults to NULL.}

\item{table_notes}{String. Note to be put under table. Not dynamic, only one note allowed. If star_substring is NULL, a caption is placed under the table. Otherwise, the comment is placed as a footnote. Passed to formatting.xtable. Defaults to NULL.}

\item{table_data.}{Data frame. No default.}
}
\description{
This function inputs table data and saves to .tex file as xtable.
}
