% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelGerdin.R
\name{ModelGerdin}
\alias{ModelGerdin}
\title{Create the Gerdin et al. model function}
\usage{
ModelGerdin(data)
}
\arguments{
\item{data}{The study data frame. No default.}
}
\description{
This function makes predictions with the Gerdin et al. model.
}
