% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelKTS.R
\name{ModelKTS}
\alias{ModelKTS}
\title{Model KTS function}
\usage{
ModelKTS(data)
}
\arguments{
\item{data}{The study data frame. No default.}
}
\description{
This function makes predictions with the Kampala Trauma Score (KTS) model.
}
