% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareStudyData.R
\name{PrepareStudyData}
\alias{PrepareStudyData}
\title{Prepare study data function}
\usage{
PrepareStudyData(study_data, data_dictionary)
}
\arguments{
\item{study_data}{The study data frame. No default.}

\item{data_dictionary}{The data dictionary object. No default.}
}
\description{
Prepares the study data using the data dictionary
}
