% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareDataForAnalysis.R
\name{PrepDataForAnalysis}
\alias{PrepDataForAnalysis}
\title{Prepare data function}
\usage{
PrepDataForAnalysis(study_data, outcome = "composite")
}
\arguments{
\item{study_data}{The study data as a data frame. No default}

\item{outcome}{The outcome variable as a string. Default: "s30d".}
}
\description{
This function prepares and generates training and review sets by splitting the dataset in half according to a time variable . Factor variables are removed.
}
