% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetUpData.R
\name{SetUpData}
\alias{SetUpData}
\title{Set up study data function}
\usage{
SetUpData(raw_study_data)
}
\arguments{
\item{raw_study_data}{imported data with no manipulations}
}
\description{
Set up study data function
}
