% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplitData.R
\name{SplitData}
\alias{SplitData}
\title{Split data function}
\usage{
SplitData(cc_df, for_tbl_one = FALSE)
}
\arguments{
\item{cc_df}{Data frame. The complete case data. No default.}

\item{for_tbl_one}{Logic. If TRUE, merged table is return with strata variable "set". Defaults to TRUE.}
}
\description{
This function sets the data for analysis, i.e. splits into to two datasets for training and test.
}
