% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToDummyVariables.R
\name{ToDummyVariables}
\alias{ToDummyVariables}
\title{To dummy variables}
\usage{
ToDummyVariables(study.sample, not.feature.variables = c("composite",
  "tc", "s"))
}
\arguments{
\item{study.sample}{Type. Description. Default/No default.}
}
\description{
This function transforms factor variables to dummy variables.
}
